Refactor smoke simulation scripts, introduce a sample configuration file for easier setup.
